Add validateUser and restricted middlewares
In this commit, I created a file called middleware inside the middlewares folder.
This file exports all middlewares to be used in all api endpoints.
I added validateNewUser() and restricted() middleware functions for validating a new user which will be registered and to restrict access to all protected resource endpoints respectively